move Editor method from Page model to PageEditor
