+++ OneMore/Commands/Search/SearchWebCommand.cs
@@ -1,11 +1,12 @@
 ﻿//************************************************************************************************
-// Copyright © 2020 Steven M Cohn.  All rights reserved.
+// Copyright © 2020 Steven M Cohn. All rights reserved.
 //************************************************************************************************
-
-using System.Threading.Tasks;
 
 namespace River.OneMoreAddIn.Commands
 {
+	using System.Threading.Tasks;
+
+
 	internal class SearchWebCommand : Command
 	{
 		public SearchWebCommand()
@@ -19,7 +20,7 @@ namespace River.OneMoreAddIn.Commands
 			string text;
 
 			await using var one = new OneNote(out var page, out _);
-			text = page.GetSelectedText();
+			text = new Models.PageEditor(page).GetSelectedText();
 
 			if (text.Length > 0)
 			{
